# openapi.yaml
openapi: 3.0.0
info:
  title: germanzero
  version: '1.0'
  license:
    name: MIT
  description: An API to access GermanZero
servers:
  - url: 'http://localhost:3000'
    description: 'Local development'
  - url: 'https://api.gz.dilab.co'
    description: 'Live server'

paths:
  /contacts:
    post:
      summary: 'Create a new contact in CRM'
      responses:
        '202':
          description: Contact is created but must be confirmed
        '409':
          description: A contact with the given email address already exist
      description: Create a new GermanZero contact and sends an email to the given address with a confirmation link
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/member'
            examples:
              example:
                value:
                  firstName: John
                  lastName: Doe
                  email: johndoe@example.com

  /contacts/:contactId/confirmations/:code:
    get:
      summary: 'Marks the email address of the contact as confirmed'
      description: Contacts are firstly created opted out from mailings to comply with GDPR. A first email is sent then, containing a confirmation link which is this route.
      responses:
        '302':
          description: Contact is confirmed, redirect to /contact-confirmed page. In case of an invalid code, the contact is not confirmed and the redirection goes to /invalid-confirmation page

  /session:
    post:
      description: Create a new session by logging in with userid and password
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                userid:
                  type: string
                password:
                  type: string
              required:
                - userid
                - password
      responses:
        200:
          description: Login successful
        403:
          description: Login failed

  /members:
    post:
      summary: Create a new GermanZero member
      responses:
        '201':
          description: Created
        '409':
          description: Conflict
      description: Create a new GermanZero member. This can only be done by someone who is already a member and therefore has a bearer token from creating a session
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/member'
                - type: object
                  properties:
                    password:
                      type: string
                      description: Initial password
            examples:
              example:
                value:
                  firstName: John
                  lastName: Doe
                  email: johndoe@example.com
                  password: my-great-secret

components:
  schemas:
    member:
      title: member
      type: object
      description: Representation of a GermanZero member record
      properties:
        id:
          type: string
          description: Automatic generated identifier for member
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
      required:
        - firstName
        - lastName
        - email
      x-examples:
        example:
          value:
            firstName: John
            lastName: Doe
            email: johndoe@example.com

  securitySchemes:
    APIKey:
      name: API Key
      type: apiKey
      in: header
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT